Chores add some logger to mp and dsm workflows

+++ src/asp_dsm.py
@@ -41,6 +41,7 @@ logger = logging.getLogger(__name__)
 
 
 def dsm_generation(params: dict, debug=False):
+    logger.info("Beggining DSM Generation")
     output_dir = params.get("output", ".")
     sources = get_sources(params)
     pairs = get_pairs(params, ids_from_source(sources))
@@ -65,26 +66,32 @@ def dsm_generation(params: dict, debug=False):
         frag += PREF_STEREO
         fragment.append(frag)
 
+    logger.info("working with {} fragments".format(len(fragment)))
+
     if "stereo" in params.keys():
         run_stereo(pairs, sources, fragment, params, debug=debug)
 
     if "pc-align" in params.keys():
+        logger.info("align fragments")
         for f in fragment:
             pc_align(
                 params["dem"], f + "-pc.tif", f + "-pc_aligned.tif", params, debug=debug
             )
 
     if "point2dem" in params.keys():
+        logger.info("rasterize fragments")
         for f in fragment:
             point2dem(f + "-pc_aligned.tif", f + "-dem.tif", params, debug=debug)
 
     if "dem-mosaic" in params.keys():
+        logger.info("merge fragments")
         dems = [f + "-dem.tif" for f in fragment]
         output = output_dir + "/dem.tif"
         dem_mosaic(dems, output, params, debug=debug)
 
 
 def run_stereo(pairs, sources, fragment, params, debug):
+    logger.info("Do stereo")
     for i, p in enumerate(pairs):
         id1, id2 = p[0], p[1]
         src1, src2 = source_from_id(id1, sources), source_from_id(id2, sources)
@@ -102,17 +109,24 @@ def run_stereo(pairs, sources, fragment, params, debug):
         stereo(mps, cams, fragment[i], params, debug=debug, dem=params["dem"])
 
 
-def retrieve_dem(params: dict) -> str:
+def retrieve_dem(params: dict, debug=False) -> str:
     bbox = retrieve_max2p_bbox(params)
     long1, long2, lat1, lat2 = bbox[0], bbox[1], bbox[2], bbox[3]
     output = params.get("output", ".")
     dst = os.path.join(output, "cop_dem30_{}_{}_{}_{}".format(int(long1), int(long2), int(lat1), int(lat2)))
-    sh("my_getDemFile.py -s COP_DEM --bbox={},{},{},{} -c /data/ARCHIVES/DEM/COP-DEM_GLO-30-DTED/DEM".format(long1, long2, lat1, lat2))
-    sh("gdal_translate -of Gtiff {} {}".format(dst + ".dem", dst + ".tif"))
 
-    os.remove(dst + ".dem")
-    os.remove(dst + ".dem.aux.xml")
-    os.remove(dst + ".dem.rsc")
+    cmd1 = "my_getDemFile.py -s COP_DEM --bbox={},{},{},{} -c /data/ARCHIVES/DEM/COP-DEM_GLO-30-DTED/DEM".format(long1, long2, lat1, lat2)
+    cmd2 = "gdal_translate -of Gtiff {} {}".format(dst + ".dem", dst + ".tif")
+    if not debug:
+        sh(cmd1)
+        sh(cmd2)
+
+        os.remove(dst + ".dem")
+        os.remove(dst + ".dem.aux.xml")
+        os.remove(dst + ".dem.rsc")
+    else:
+        print(cmd1)
+        print(cmd2)
     return dst + ".tif"
